Yarn Dedup & rename hasMoreThanOneBinding to shouldStoreRHSInTemporaryVariable
