+++ packages/babel-plugin-proposal-object-rest-spread/package.json
@@ -28,8 +28,8 @@
   },
   "devDependencies": {
     "@babel/core": "workspace:*",
-    "@babel/parser": "workspace:*",
-    "@babel/helper-plugin-test-runner": "workspace:*"
+    "@babel/helper-plugin-test-runner": "workspace:*",
+    "@babel/parser": "workspace:*"
   },
   "engines": {
     "node": ">=6.9.0"
